+++ CHANGELOG.md
@@ -687,3 +687,4 @@
 * Change precedence for `<-*-*-` and `.-*-*-`
 * Change precedence for `<-|-<-|-`, `<-|->-|-`, `>-|->-|-`, `>-|-<-|-`
 * Define length encoding operators from `<-*--` to `<-*-------`
+* Define `<-||-` and `>-||-` operators in `Algebraic` module